Relocating the access token manager.
