+++ src/services/automower/getMowersService.ts
@@ -1,4 +1,4 @@
-import { AccessTokenManager } from '../authentication/accessTokenManager';
+import { AccessTokenManager } from './accessTokenManager';
 import { AutomowerClient } from '../../clients/automowerClient';
 import { NotAuthorizedError } from '../../errors/notAuthorizedError';
 import { Mower } from '../../model';
