+++ test/automowerPlatformSpy.ts
@@ -1,5 +1,5 @@
 import { AutomowerPlatform } from '../src/automowerPlatform';
-import { AccessTokenManager } from '../src/services/authentication/accessTokenManager';
+import { AccessTokenManager } from '../src/services/automower/accessTokenManager';
 import { EventStreamService } from '../src/services/automower/eventStreamService';
 import { DiscoveryService } from '../src/services/automower/discoveryService';
