+++ test/services/automower/mowerControlService.spec.ts
@@ -2,7 +2,7 @@ import { It, Mock, Times } from 'moq.ts';
 import { AutomowerClient } from '../../../src/clients/automowerClient';
 import { NotAuthorizedError } from '../../../src/errors/notAuthorizedError';
 import { AccessToken } from '../../../src/model';
-import { AccessTokenManager } from '../../../src/services/authentication/accessTokenManager';
+import { AccessTokenManager } from '../../../src/services/automower/accessTokenManager';
 import { Action, MowerControlServiceImpl } from '../../../src/services/automower/mowerControlService';
 
 describe('MowerControlService', () => {
